- replace shouldNotAddWrapperElement and applyWrapperElementToSiblings with one helper method - wrapBlockNodes
